[BEAM-12830] Print clearer go version fail message

+++ sdks/go/prepare_go_version.sh
@@ -34,7 +34,7 @@ GOVERS="invalid"
 
 if ! command -v go &> /dev/null
 then
-    echo "go could not be found. This script requires a go installation > 1.16 to bootstrap using specific go versions."
+    echo "go could not be found. This script requires a go installation > 1.16 to bootstrap using specific go versions. See http://go.dev/doc/install for options."
     exit 1
 fi
 
@@ -54,13 +54,27 @@ case $key in
 esac
 done
 
+# Versions of Go > 1.16 can download arbitrary versions of go.
+# We take advantage of this to avoid changing the user's
+# installation of go ourselves. 
+MINGOVERSION="go1.16.0"
+
+# Compare the go version by sorting only by the version string, getting the
+# oldest version, and checking if it contains "min". When it doesn't, it's
+# the go print out, and it means the system version is later than the minimum.
+if (echo "min version $MINGOVERSION os/arch"; go version) | sort -Vk3 -s |tail -1 | grep -q min;
+then 
+  # Outputing the system Go version for debugging purposes.
+  echo "System Go installation at `which go` is `go version`, is older than the minimum required for hermetic Beam builds. Want $MINGOVERSION. See http://go.dev/doc/install for installation instructions."; 
+  exit 1
+fi
+
 GOPATH=`go env GOPATH`
 GOBIN=$GOPATH/bin
 GOHOSTOS=`go env GOHOSTOS`
 GOHOSTARCH=`go env GOHOSTARCH`
 
-# Outputing the system Go version for debugging purposes.
-echo "System Go installation: `which go` is `go version`; Preparing to use $GOBIN/$GOVERS"
+echo "Preparing to use $GOBIN/$GOVERS"
 
 # Ensure it's installed in the GOBIN directory, using the local host platform.
 GOOS=$GOHOSTOS GOARCH=$GOHOSTARCH GOBIN=$GOBIN go install golang.org/dl/$GOVERS@latest
